+++ src/components/Header/header.module.scss
@@ -1,0 +1,41 @@
+@import '../../sass/main.scss';
+
+.header {
+  // background: $color-purple-dark;
+  margin-bottom: 1.45rem;
+  text-align: center;
+  position: relative;
+
+  & a,
+  a:link,
+  a:visited {
+    color: $color-purple-dark;
+    text-decoration: none;
+    display: inline-block;
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    box-shadow: 0px 0px 100px 90px $color-white;
+    background-color: $color-white;
+
+    z-index: 100;
+    &:hover {
+      color: $color-orange-light;
+    }
+
+    @include respond('tab-port') {
+      box-shadow: 0px 0px 60px 70px $color-white;
+    }
+  }
+}
+
+.heading {
+  font-family: 'Lilita One', cursive;
+  font-size: 8.4rem;
+  letter-spacing: 0.2rem;
+}
+
+.subheading {
+  font-size: 2rem;
+}
